+++ bika/lims/browser/referencesample.py
@@ -5,21 +5,25 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-import json
+import collections
 from datetime import datetime
-from operator import itemgetter
 
-from Products.ATContentTypes.utils import DT2dt
-from Products.Archetypes.config import REFERENCE_CATALOG
-from Products.CMFCore.utils import getToolByName
-from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from bika.lims import bikaMessageFactory as _, logger
+from bika.lims import api
+from bika.lims import bikaMessageFactory as _
+from bika.lims import logger
 from bika.lims.browser import BrowserView
 from bika.lims.browser.analyses import AnalysesView
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.browser.chart.analyses import EvolutionChart
+from bika.lims.utils import get_image
+from bika.lims.utils import get_link
 from bika.lims.utils import t
 from plone.app.layout.globals.interfaces import IViewView
+from plone.memoize import view
+from Products.Archetypes.config import REFERENCE_CATALOG
+from Products.ATContentTypes.utils import DT2dt
+from Products.CMFCore.utils import getToolByName
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope.interface import implements
 
 
@@ -190,78 +194,132 @@ class ReferenceAnalysesView(AnalysesView):
 
 
 class ReferenceResultsView(BikaListingView):
+    """Listing of all reference results
     """
-    """
+
     def __init__(self, context, request):
         super(ReferenceResultsView, self).__init__(context, request)
-        self.title = self.context.translate(_("Reference Values"))
-        self.description = self.context.translate(_(
-            "Click on Analysis Categories (against shaded background) "
-            "to see Analysis Services in each category. Enter minimum "
-            "and maximum values to indicate a valid results range. "
-            "Any result outside this range will raise an alert. "
-            "The % Error field allows for an % uncertainty to be "
-            "considered when evaluating results against minimum and "
-            "maximum values. A result out of range but still in range "
-            "if the % error is taken into consideration, will raise a "
-            "less severe alert."))
-        self.contentFilter = {}
+
+        self.catalog = "bika_setup_catalog"
+        self.contentFilter = {
+            "portal_type": "AnalysisService",
+            "UID": self.get_reference_results().keys(),
+            "inactive_state": "active",
+            "sort_on": "sortable_title",
+            "sort_order": "ascending",
+        }
         self.context_actions = {}
+        self.title = self.context.translate(_("Reference Values"))
+        self.icon = "{}/{}".format(
+            self.portal_url,
+            "/++resource++bika.lims.images/referencesample_big.png"
+        )
         self.show_sort_column = False
         self.show_select_row = False
         self.show_workflow_action_buttons = False
         self.show_select_column = False
         self.pagesize = 999999
+        self.show_search = False
 
-        self.columns = {
-            'Service': {'title': _('Service')},
-            'result': {'title': _('Result')},
-            'min': {'title': _('Min')},
-            'max': {'title': _('Max')},
-        }
+        # Categories
+        if self.show_categories_enabled():
+            self.categories = []
+            self.show_categories = True
+            self.expand_all_categories = True
+            self.category_index = "getCategoryTitle"
+
+        self.columns = collections.OrderedDict((
+            ("Title", {
+                "title": _("Analysis Service"),
+                "sortable": False}),
+            ("result", {
+                "title": _("Expected Result"),
+                "sortable": False}),
+            ("error", {
+                "title": _("Permitted Error %"),
+                "sortable": False}),
+            ("min", {
+                "title": _("Min"),
+                "sortable": False}),
+            ("max", {
+                "title": _("Max"),
+                "sortable": False}),
+        ))
+
         self.review_states = [
-            {'id': 'default',
-             'title': _('All'),
-             'contentFilter': {},
-             'columns': ['Service',
-                         'result',
-                         'min',
-                         'max']},
+            {
+                "id": "default",
+                "title": _("All"),
+                "contentFilter": {},
+                "columns": self.columns.keys()
+            }
         ]
 
-    def folderitems(self):
-        items = []
-        uc = getToolByName(self.context, 'uid_catalog')
-        # not using <self.contentsMethod=bsc>
-        for x in self.context.getReferenceResults():
-            service = uc(UID=x['uid'])[0].getObject()
-            item = {
-                'obj': self.context,
-                'id': x['uid'],
-                'uid': x['uid'],
-                'result': x['result'],
-                'min': x['min'],
-                'max': x['max'],
-                'title': service.Title(),
-                'Service': service.Title(),
-                'type_class': 'contenttype-ReferenceResult',
-                'url': service.absolute_url(),
-                'relative_url': service.absolute_url(),
-                'view_url': self.context.absolute_url() + "/results",
-                'replace': {},
-                'before': {},
-                'after': {},
-                'choices': {},
-                'class': {},
-                'state_class': 'state-active',
-                'allow_edit': [],
-            }
-            item['replace']['Service'] = "<a href='%s'>%s</a>" % \
-                (service.absolute_url(), service.Title())
-            items.append(item)
+    def update(self):
+        """Update hook
+        """
+        super(ReferenceResultsView, self).update()
+        self.categories.sort()
 
-        items = sorted(items, key=itemgetter('Service'))
-        return items
+    @view.memoize
+    def show_categories_enabled(self):
+        """Check in the setup if categories are enabled
+        """
+        return self.context.bika_setup.getCategoriseAnalysisServices()
+
+    @view.memoize
+    def get_reference_results(self):
+        """Return a mapping of Analysis Service -> Reference Results
+        """
+        referenceresults = self.context.getReferenceResults()
+        return dict(map(lambda rr: (rr.get("uid"), rr), referenceresults))
+
+    def folderitem(self, obj, item, index):
+        """Service triggered each time an item is iterated in folderitems.
+
+        The use of this service prevents the extra-loops in child objects.
+
+        :obj: the instance of the class to be foldered
+        :item: dict containing the properties of the object to be used by
+            the template
+        :index: current index of the item
+        """
+
+        uid = api.get_uid(obj)
+        url = api.get_url(obj)
+        title = api.get_title(obj)
+
+        # get the category
+        if self.show_categories_enabled():
+            category = obj.getCategoryTitle()
+            if category not in self.categories:
+                self.categories.append(category)
+            item["category"] = category
+
+        ref_results = self.get_reference_results()
+        ref_result = ref_results.get(uid)
+
+        item["Title"] = title
+        item["replace"]["Title"] = get_link(url, value=title)
+        item["result"] = ref_result.get("result")
+        item["min"] = ref_result.get("min")
+        item["max"] = ref_result.get("max")
+
+        # Icons
+        after_icons = ""
+        if obj.getAccredited():
+            after_icons += get_image(
+                "accredited.png", title=_("Accredited"))
+        if obj.getAttachmentOption() == "r":
+            after_icons += get_image(
+                "attach_reqd.png", title=_("Attachment required"))
+        if obj.getAttachmentOption() == "n":
+            after_icons += get_image(
+                "attach_no.png", title=_("Attachment not permitted"))
+        if after_icons:
+            item["after"]["Title"] = after_icons
+
+        return item
 
 
 class ReferenceSamplesView(BikaListingView):