+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
 **Changed**
 
 - #1091 New Worksheet blank/control/duplicate listings
+- #1093 Listing integration for Analysis Specification Widget
+- #1092 Listing integration for Profile Analyses Widget
 - #1081 API functions improved
 - #1076 Instrument QC Viewlet Availability
 - #1071 Reinvented Listing Tables